refactor: 프로젝트 구조 변경
- domain 안의 패키지들을 상위 패키지로 이동함.
