+++ src/test/java/lotto/domain/vo/lottoresult/LottoRankTest.java
@@ -3,6 +3,7 @@ package lotto.domain.vo.lottoresult;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.stream.Stream;
+import lotto.vo.lottoresult.LottoRank;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
