+++ src/test/java/lotto/service/lotto/LottoResultGeneratorTest.java
@@ -5,12 +5,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
-import lotto.domain.vo.Budget;
-import lotto.domain.vo.DrawnLottoNumber;
-import lotto.domain.vo.LottoCount;
-import lotto.domain.vo.lotto.Lotto;
-import lotto.domain.vo.lotto.Lottos;
-import lotto.domain.vo.lottoresult.LottoResults;
+import lotto.vo.Budget;
+import lotto.vo.DrawnLottoNumber;
+import lotto.vo.LottoCount;
+import lotto.vo.lotto.Lotto;
+import lotto.vo.lotto.Lottos;
+import lotto.vo.lottoresult.LottoResults;
+import lotto.service.LottoResultGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
